♻️ refactor: getCookieMap util로 분리시킴
