+++ src/common/utils/cookie.ts
@@ -1,0 +1,11 @@
+export const getCookieMap = () => {
+  const cookieArray: string[][] = document.cookie
+    .split('; ')
+    .map((entry) => entry.split('='));
+
+  const iterableCookieArray: Array<readonly [string, string]> = cookieArray.map(
+    ([key, value]) => [key, value]
+  );
+
+  return new Map(iterableCookieArray);
+};
